services/irrigation-service/requirements.txt
fastapi==0.115.6
pydantic==2.10.3
paho-mqtt==2.1.0
requests==2.32.3
redis==5.2.1
python-dotenv==1.0.1

pytest==8.3.4
pytest-cov==6.0.0
psycopg2-binary==2.9.10
